requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argon2==0.1.10
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
flask-cors==3.0.10
flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
jinja2==3.1.2
libgravatar==1.0.0
markupsafe==2.1.1
nanoid==2.0.0
packaging==21.3; python_version >= '3.6'
passlib==1.7.4
pluggy==1.0.0; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyjwt==2.4.0
pymongo==4.1.1
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest==7.1.2
python-arango==7.3.4
python-dotenv==0.20.0
requests-toolbelt==0.9.1
requests==2.27.1
setuptools==62.4.0; python_version >= '3.7'
six==1.16.0
tomli==2.0.1; python_version >= '3.7'
urllib3==1.26.9
uwsgi==2.0.20
werkzeug==2.1.2
zipp==3.8.0
